backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14
django-cors-headers>=4.3
Pillow>=10.0
python-dotenv>=1.0
gunicorn>=21.2
# Cloudflare R2 (через boto3 + django-storages)
boto3>=1.34
django-storages>=1.14
drf-spectacular>=0.27
django-filter>=24.2
dj-database-url>=2.1
psycopg2-binary>=2.9
